⭐ make executor keep track of score and query results
this is the first step in a bit of a larger overhaul. right now every query can give us multiple results, but when we did the scoring we only looked at the last entrypoint, i.e. only one of the entrypoints in case multiple were given. this was never much of an issue with blocks that had multiple entrypoints, since the ended up being just one entrypoint on the outermost layer and aggregated their result value. bot it is an issue for the scoring now, which can miss entire results.

this first step is only about adding a separate scoring:
1. code bundles now carry their own ID and it is separately calculated, i.e. it is never equal to any entrypoint ID
2. code bundle IDs are calculated from all entrypoints (and called blocks)
3. executor keeps track of all entrypoint results and once it has all entrypoints for a code bundle collected, it will return a score for the overall query with the bundle ID