✨🧹 correctly expose parser errors
we are currently eating up parser errors in favor of compiler errors, to help let the compiler give us hints about what could be improved.

this is causing problems in cases when the underlying parsing syntax is wrong, and leads to really weird error messages. Latest example: After introducing switch-case statements, i had added the `:` operand in the case statement. However, the resulting error I got had nothing to do with it anymore, saying that the child block is missing. In reality the parser just didnt understand the `:` when encountering it.

to fix it, the parser error messages had to be improved significantly [done]

also we are keeping the compile step because the bundle can still generate helpful suggestions, which are provided to the user. however unlike before, if there is a parser error, we always return that first.

[bonus]: the parser gives us the location in its error too, which is awesome.