🌩️  Support JSON logging for nexus
Currently, the kubernetes setup is using stackdriver logging. However,
given that we are in GKE, I think we shouldn't need this and can just
output json formatted logs. The GKE cluster is already running
fluentbit, which will take care of pushing the logs to google logging:
https://cloud.google.com/stackdriver/docs/solutions/gke/managing-logs

The benefit of logging to stdout/stderr is that we can see the logs for
each pod either through kubectl/k9s or the GKE dashboard directly.

Here's what the config looks like:
```
logging:
  level: debug
  writer: cli
  options:
    format: gcp-json
```