🐛 fix any on null lists
Calling any on a `null` list results in the internal compilation into `nullList.length != 0`. With null-chaining this results in `null != 0` which returns true and thus the call succeeds, even though it should not. We introduce a solution here where we check for non-null conditions as well as existing list contents. An internal call is introduced to avoid compiler workarounds.