🐛 wire through user specified platform identifiers to NewMotorConnection
User specified id detectors were not respected.

We are resolving the asset twice. The first time, the asset was being
resolved with the default platform identifiers, and then being Sync'd.
The second time, the user specified id detectors were being used. This
was problematic in #529 because the only platform identifier that was to
be used was the instance identifier, but the first time, because from
mondoo's perspective it was a local scan, we used hostname and
machineid.

Fixes #529