🐛 fix mondoo cli output
I was seeing that the logging was not being correctly paused. We changed
some code to use logger.FromContext, which also ended up being called in
the mondoo cli path. If a request logger is not set, that code returned
a global logger which was created before the mondoo cli overrode the log
output writer, which meant that it didn't get the message to pause