🐛 make a copy of the iterator value in arrayWhere
This was caught by the race detector in #628. After changing this,
2/5 are fixed. The bug here is that closure captures the i form the
for block. It's not a copy, but a reference. I'm guessing sometimes,
runFunctionBlock becomes asynchronous and returns before the result is
complete. If that happens, the i value you'd get will probably the the
last one in the list