🐛 make files permission id unique
This is part of the work required to get this query working:
```
files.find(from: "/usr/bin", xdev: false).list.where(permissions.suid)
```

This addresses an issue where for each list item, we create a new
executor. However, that executor shares the lumi cache. I was ending up
with a situation where something was being found in the cache that
didn't match exactly, and thus the wrong result was being fed through
the chain

fixes #654