🐛 Blocks that produce single values should have 1 entrypoint
Consider a system where `/etc/resolv.conf` is not empty. The following
query should succeed:
```js
if (true) {
  a = file("/etc/resolv.conf").content != ""
  b = false
  return a || b
}
```

However, that was not happening. Instead, the result was usually coming
back as failed. Sometimes, like when in a debugger, it was possible for
it to succeed. Further analysis of the compiled code showed that 3
entrypoints.

Code the processes blocks assume single valued blocks only produce one
result. And so the first of these that was received was being set to the
result.