🫤 Make files.find follow symlinks
We have inconsistencies between our filesystem implementations and
files.find. By default, mose of filesystem implemenations seem to be using
"stat -L", which follows symlinks. They don't implement a way to read
symlinks. The one exception I see is we use afero.NewOsFs for local
transports.

Doing this should fix #701. I'm not sure this is a good solution. I
think we need to make it explicit how symlinks are handled and make sure
all our implementations conform to that