🧹 Refactor the platform id detection
With this PR, the platform id detection now happens as part of the
discovery phase instead of being tied to the motor creation. One
tangible benefit right now is that it no longer runs the detection code
multiple times, as we create the motor connections multiple times.

Also changing is the way the `mondoo.asset.platformIDs` function works.
Because we no longer have the platform ids on the motor, we need to
inject this information into the runtime.