🐛 return applicable callbacks from triggered chains
This is another edge-case which can be provoked with the upcoming list-block assertions, which run the assertion (all/none/one/any) on top of a list, which both need to be returned. When the list is initially generated, it calls triggerChain but does not itself return. With bad timing, this happens after the outer runner has visited this entrypoint without having a value. May generate unnecessary trigger calls, but some are required.