⭐ array assertion `@msg` to support `$expected`
This also required a pretty heavy change in the way functions are called
and compiled. This only works for `${}` blocks for now, but can quickly
be extended to others if we need it. It allows us to bind variables into
the block that otherwise wouldn't exist. That said, I'm still pondering
a different way to communicate vars across which might impact this as
well... (ie paying the piper)