✨ change the type of `if` to block (unless it returns)
The type of `if` was the type of the last expression in the block. This
changes it to be either a block, so we collect all the results, or in
the case of using `return`, whatever the type of that statement is.

We need to collect the block to provide useful information.