🐛 cache error in runResourceFunction
There were errors being returned here that were not being put in the
cache. Those errors happened on codepoints that were not entrypoints or
datapoints, and thus end up being ignored by the main execution loop.
These need to be stored as error so that the datapoint and entrypoints
load the errors and report themselves as errored. Otherwise, the
entrypoints and datapoints never report and cause execution to never
complete

Partial fix for #957