⭐️ Piper code
This change restructures the compiled MQL code to allow for referencing across scopes. This currently requires using the feature flag `PiperCode`.
For example, previously, the following did not work:
```
v = "unstable"
mondoo {
  version == v
}
```

This problem is solved by flattening the compiled code instead of nesting it in a tree like structure. Once flattened into blocks, blocks are able to reference previous blocks. This is done by expanding references to be 64 bit values. The top 32 bits are the block id, and the lower are the chunk id. The example above would compile into something like:
```
<block 1, chunk 1>: primitive("unstable")
<block 1, chunk 2>: resource("mondoo")
<block 1, chunk 3>: call("{}", bind(ref<block 1, chunk 2>), args=(), depends_on(ref<block 1, chunk 1>))

<block 2, chunk 1>:  context
<block 2, chunk 2>: call("version", bind(ref<block 2, chunk 1>), args=(), depends_on())
<block 2, chunk 3>: primitive(ref<block 1, chunk 1>)
<block 2, chunk 4>: call("==", bind(ref<block 2, chunk 2>), args=(<block 2, chunk 3>), depends_on())
```

Note that `depends_on` is actually just part of args in the implementation. Dependencies are appended to the arguments to force the dependency. Each block as a parameters value that keeps track of the number of expected paramaters (args). Anything more than this is not passed to the block, but is added as a dependency.

To see the actual generated code for the previous example:
```
FEATURES=PiperCode go run apps/mondoo/mondoo.go exec --ast 'v = "unstable"
                                 mondoo {
                                   version == v
                                 }'

{
  "code_v2": {
    "blocks": [
      {
        "chunks": [
          {
            "primitive": {
              "type": "\u0007",
              "value": "dW5zdGFibGU="
            }
          },
          {
            "call": 1,
            "id": "mondoo"
          },
          {
            "call": 1,
            "function": {
              "args": [
                {
                  "type": "\u001c\u0000",
                  "value": "gICAgEA="
                },
                {
                  "type": "\u0003",
                  "value": "goCAgCA="
                }
              ],
              "binding": 4294967298,
              "type": "\u000c"
            },
            "id": "{}"
          }
        ],
        "entrypoints": [
          4294967299
        ]
      },
      {
        "chunks": [
          {
            "primitive": {
              "type": "\u001bmondoo"
            }
          },
          {
            "call": 1,
            "function": {
              "binding": 8589934593,
              "type": "\u0007"
            },
            "id": "version"
          },
          {
            "primitive": {
              "type": "\u0003",
              "value": "goCAgCA="
            }
          },
          {
            "call": 1,
            "function": {
              "args": [
                {
                  "type": "\u0003",
                  "value": "hoCAgEA="
                }
              ],
              "binding": 8589934594,
              "type": "\u0004"
            },
            "id": "==\u0007"
          }
        ],
        "entrypoints": [
          8589934596
        ],
        "parameters": 1
      }
    ],
    "checksums": {
      "4294967297": "/zmPyM1RiG0+q2P8rsxw2UYCjZ1qXOnmWbO2qiqEQVjnkc+4OYSrtX/y5/x3TfurzJlu0HN2n+wxLHy8skzZJw==",
      "4294967298": "us0KibGykMCbdBhLtA3fC6F5hcvNz1vZRfFIPcit53W9d35n+P2AvhRrWgzjFxqRfB8VlFW4nwBeFDu4gzkDQA==",
      "4294967299": "dxKYbmxTm6OcULtinXezDhH44hJrDV4fsNy0Wl4NuxgjV5yDgGVBLXPv8GzYeAsjHeiF+ctIPWZHFkx2QkPXrA==",
      "8589934593": "us0KibGykMCbdBhLtA3fC6F5hcvNz1vZRfFIPcit53W9d35n+P2AvhRrWgzjFxqRfB8VlFW4nwBeFDu4gzkDQA==",
      "8589934594": "J4anmJ+mXJX380Qslh563U7Bs5d6fiD2ghVxV9knAU0iy/P+IVNZsDhBbCmbpJch3Tm0NliAMiaY47lmw887Jw==",
      "8589934595": "gEYZ80qjDUAoxWxJJ6i+mPsv+qO1w8NzbCENi34HQbjXOkzWIVRsI/dXT/4fPWefdjwjsQOkVadCoJuOB4NjFQ==",
      "8589934596": "2sESxSBAffJ5nFmLSaRJoauQwKlivisuy6eoduaY4kuZkiJ1QAHPu9O5o+L4zbsNm4Q/U5AmSdB5j4/7wwqHTA=="
    },
    "id": "pG3hKFdYgbA="
  },
  "labels": {
    "labels": {
      "2sESxSBAffJ5nFmLSaRJoauQwKlivisuy6eoduaY4kuZkiJ1QAHPu9O5o+L4zbsNm4Q/U5AmSdB5j4/7wwqHTA==": "version == <ref>",
      "dxKYbmxTm6OcULtinXezDhH44hJrDV4fsNy0Wl4NuxgjV5yDgGVBLXPv8GzYeAsjHeiF+ctIPWZHFkx2QkPXrA==": "mondoo"
    }
  },
  "min_mondoo_version": "5.15.0",
  "source": "v = \"unstable\"\nmondoo {\n  version == v\n}",
  "version": "unstable"
}
-> block 1
   entrypoints: [<1,3>]
   1: "unstable"
   2: mondoo
   3: {} bind: <1,2> type:block (=> <2,0>, ref<1,1>)
-> block 2
   entrypoints: [<2,4>]
   1: mondoo id = context
   2: version bind: <2,1> type:string
   3: ref<1,1>
   4: == bind: <2,2> type:bool (ref<2,3>)
```

This change affects policy resolution. Any queries compiled by the server now always compile both versions.
The code id from the piper query is always used as it is expected to compile strictly more things.
If the client says its using the PiperCode feature, during policy resolution, we make sure to use the
piper code to generate the resolved policy. This involves making sure the checksums for the collector job
and execution job are unique for piper code resolved policies so as not to interfere with non piper code resolved
policies. We also now store additional metadata on the asset that says it resolved a policy with the PiperCode
feature ("v2" resolved policy) to allow us to generate the reports correctly