🧹 make binaries smaller
This moves the `GraphBetaClient` to the resources package where it is
being used. For some reason, when it is in motor, go is not eliding a
bunch of unused code from the `msgraphbetasdk` package.

Here are some statistics:
```
$ cat notes/new_lambda | wc -l
211444
$ cat notes/old_lambda | wc -l
371471
$ cat notes/old_lambda | grep msgraph | wc -l
187075
$ cat notes/new_lambda | grep msgraph | wc -l
27047
$ cat notes/new_lambda | grep -v msgraph | wc -l
184397
$ cat notes/old_lambda | grep -v msgraph | wc -l
184396
```

There's about 160K additional symbols in the old version, all but one
seems to come from msgraph. I'm guessing the other one is from the
method removed in motor