🐛 fix flag parsing
We were running into an issue where the flag parser was not handling default expectations correct. It was not able to deal with arguments and --key value chains.

It is difficult to generically parse flags properly since a --flag could be a string flag which is followed by a value or a boolean flag which is not followed by a value.

Therefore we can never provide a perfect solution for the case but we can make it better than current regex-based methods. Therefore we go with most sensible defaults.

Therefore we are going to handle the case when `--key value` is provided that we expect key to be a string value and not be a boolean value.