🐛 assign pkg format from pkg instead from package manager
we had the issue that different package manages may be supported on one system (e.g. windows appx and hotfixes). We loose the source of that information when we are not using the pkg object format field and instead use the information from the pkg manager. This change updates all the different information to add the pkg format field directly.