🧹 Refactor uses of runFunctionBlock in v1 code
Things that were using runFunctionBlock were not always dealing with the
quirks of that function correctly. They also had difficult to understand
closures which sometimes did correct locking. I've refactored things to
deal with those in a single function. Users of runFunctionBlock now call
a new function which is provided a callback which is called once on
completion