🐛 fix ebs scan no asset error
correct volume detection with ebs volume scanning is tricky.
aws makes no guarantees that it will place the volume where we told
it to.

this change falls back to looking for the non root volume if we can't
find the matching named volume.

for the no setup case, it further falls back to getting the root volume if
the non-root volume cannot be found. that is bc the nosetup case
assumes the target volume has already been mounted to the instance