🧹 reduce number of msgraph packages pulled in
This reduces the total packages from around 17k to 7k.

Build time went from 8min57s to 4min49s for builds with
no caching on my machine. If we can convince the ms graph
module to compile earlier, there's 30 seconds where only
1 of 12 threads are being used on my machine because its
waiting for the msgraph models package to compile before it
can compile the result of the msgraph module