🐛 Fix assertion messages broken since v6
The assertion message structure in the code bundle can produce
collisions if the checksums for a codepoint are the same. This ends up
happening for a case like:
```
  # @msg ${length} users
  users.none(uid == 0)
```

The checksum of the codepoint for `none` is used in the assertion
message, but both v5 and v6 end up with the same value. The checksums
inside the block are different.

This fixes the problem by making sure the 2 versions dont share the
same variable inside the codebundle