🐛 Make sure changing assertion messages changes the code id
The execution checksums only change when the code id changes. We seem to
be doing the right thing with the v6 code id generation, however
everything still uses v5. This should be fine as long as assertion
messages are not used in asset filters, which would likely cause
disagreement in what the code id is with old clients