⭐ simple multi-line mondoo shell
This is - by far - not something I'd consider a good implementation, but it's a first step towards multi-line, which has been annyoing in shell for a long time.

Why? Because in this implementation you cannot go back to previous lines and edit them. We don't have a nice editor that can plug closing brackets when you open something up or allow us to type in the middle of an expression and break it over multiple lines. To my knowledge, if we want to do that, we cannot use go-prompt anymore, since it doesn't support it. This approach to multi-lines is supported in go-prompt, but it's definitely limiting.