🧹 MQL Boolean Assertions
In MQL today, when you write a control, it checks to see if the overall
result is truthy. If you look at a block, for example, it is truthy if
all its values are truthy. However, this leads to problems with blocks
like this:

```
users.list {
  uid
  name != ""
}
```

Even if all names comply with the condition, it will still fail on the
user whose uid is 0. This is because 0 is falsy and thus the block
evaluates to being false.

This change makes it so we only check boolean results when determining
if a query succeeded. In the above example, `uid` would be ignored in
that context.

This change is a breaking change, and thus behind a feature flag until
v7

To use it, specify the environment variable `FEATURE=BoolAssertions`