allow local scan with sudo
Presently, performing a 'mondoo scan local --sudo' does not actually do
anything with regards to 'sudo'. Follow a similar pattern to how the ssh
provider can make use of sudo by pre-pending commands to execute as
necessary, and re-using the 'cat' filesystem that the ssh provider uses
to access files with.

Signed-off-by: Joel Diaz <joel@mondoo.com>