⭐ lr go to generate Schema json
We can then use this schema JSON to pre-inform the compiler and other parts about the structure of all offered resources. It helps us separate execution-time from compile-time dependencies

Note: There is a CLI change that removes the `--stdout` option from
`lr`, which was not used anyway. This is because we want to generate
both the `.go` as well as the `.json` files in the same call.

It is consolidate into one call because:
1. both files are needed anyway in every run
2. it's faster
3. it's fewer things for users to do