✨ move manifest from .go to .json embed (#10)
The previous .go import moved the generation to the compiler. For now, we are going to move it to a json construct, similar to the way resource pack infos are loaded. The manifests contain a lot of useful information which should be available via other APIs, so a JSON structure comes handy.

This is also only an initial step to actually merge the two info streams together into one.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>