🐛 interpret nil bytes as empty string
if we cast an empty string to `[]byte`, and put it in a protobuf
`bytes`, that doesn't get serialized/serialized as an empty array.
It comes back as null. Which means we cannot distinguish between an
empty string and a nil string

We ran into this case with this query:
```
  registrykey.property(path: 'HKEY_LOCAL_MACHINE\SYSTEM\CurrentControlSet\Services\LanManServer\Parameters', name: 'NullSessionPipes') {
    value == ""
  }
```

It worked fine when we ran it locally with `mondoo exec`. It worked fine
when we ran in in a policy in incognito mode. It failed when running
from a policy against nexus. The resolved policy that came back had a
nil Value for the primitive intead of an empty `[]byte`

Signed-off-by: Jay Mundrawala <jay@mondoo.com>