use path instead of filepath and io instead of ioutil
'filepath' uses the OS-specific idea of path separators which would make
a mess of path cleanup of UNIX-paths when run from a Windows machine.
'path', on the other hand hard codes the idea of '/' as a path separator
so use it when we know we're dealing with UNIX-paths.

'io/ioutil' is deprecated and as of golang 1.16, 'io' has the
functionality we want, so use 'io' instead.

Signed-off-by: Joel Diaz <joel@mondoo.com>