🛑 overhaul nil-handling
Until now, every typed value could be nil. However, this is causing problems with proto serialization where the state is not deterministic and thus nil fields may turn into empty array.

With this change typed values that are set to nil are also turned into a types.Nil.

Because the compiled code comes with type information, we can turn Primitives and RawData into the Nil type while retaining knowledge about the original type they are supposed to fill.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>