✨ introduce impact to mquery (#145)
This is designed to allow users to help prioritize different queries that they are running. Impact can be used in a variety of ways, e.g. to identify important queries for security, incidence response, or operational upkeep. Higher impact values signify more improtant queries. Values are from 0 - 100. This feature is experimental at this time in cnquery.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>