⭐ explore report --json (#270)
Introduces a custom printer to create reports. It doesn't just take the
raw data and dump it all to CLI, but instead carefully constructs a JSON
representation of the report that is useful for further processing. It
contains MRNs for queries, eliminates pesky lookups, and nicely
aggregates data.

Features: 3 fields on the top level: assets (for quick lookup and
aggregation), reports (for all the data) and errors (for error
aggregation).

Unlike `json.Marshal` it doesn't dump any compiled code, doesn't dump
the resolved pack, and doesn't force lookups of all data fields.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>