⭐ cnquery scan ..
Rename the  command to , because it is more intuitive. It still explores the assets given the provided query packs, but semantically it's easier to remember to "scan an asset" like the crew in Star Trek would scan a planet they were looking at. We explore things by looking at the gathered data or walking the planet, but first we scan it.

Note: the package was not renamed, because it contains a scanner inside of it (see ). So we can think of it as the space ship that has an explorer module with a scanner that does the work.
Signed-off-by: Dominik Richter <dominik.richter@gmail.com>