🌟 default fields in resources (#303)
Only after #302 (to avoid working through a rebase).

This allows us to define default values for resources to be printed and
collected.

The simplest example is `mondoo`. When users enter it in the shell
today, they get:

```
mondoo
mondoo: mondoo id = mondoo
```

This is not helpful at all. Instead, we want to print useful information
about resources.

With this feature we can define default fields and execute them:

```
mondoo
mondoo: mondoo version="v7.0.0-alpha1"
```

This becomes especially powerful for resources with a lot of fields:

```
cnquery> users[0]
users.list[0]: user name="root" uid=0 gid=0
```

In conjunction with the list expansion that merged, we now get:

```
cnquery> users
users.list: [
  user name="root" uid=0 gid=0
  user name="bin" uid=1 gid=1
  user name="daemon" uid=2 gid=2
  user name="adm" uid=3 gid=4
```

Still needs a bit of follow-up work to make it work with the current
list assertions.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>