🐛 fix type for nil results (#315)
When converting from raw data to primitive, we have a case where the
value is `nil` and no error was given. In this case, the primitive
should represent it as `Type = types.Nil`. This is done because proto
cannot properly differentiate between a value being nil or a value being
an empty list.

Due to this, the desired behavior for primitives (not raw data!) and
results (which are based on primitives) is:

1. If it's a non-empty result: `Value = ..., Type = ...`
2. If it's an empty result: `Value = [], Type = ...`
3. If it's a nil value: `Value = nil, Type = nil`

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>
Co-authored-by: Jay Mundrawala <jay@mondoo.com>