🐛 only render progress bar in tty mode (#341)
The fixes an issue where a user that runs

```
cnquery scan local --json | jq .
```

was erroring out with:

```
cnquery scan local --json | jq .                                                                                                                                                                                
→ using service account credentials
→ discover related assets for 1 asset(s)
→ resolved assets resolved-assets=1
→ connecting to asset spacerocket.fritz.box (baremetal)
parse error: Invalid numeric literal at line 2, column 106
```

The reason was the progress bar that was written to stdout. The new
change only prints the progress bar when the stdout is attached to a
tty.