🌈 add query title or UID to output
It will try to grab either to print it. If neither is available, we will just stick to printing the output of the query itself. It will improve over time, so that it becomes clearer where it originated. At least for the short version, we want to avoid printing the full query.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>