🌈 add query title or UID to output (#345)
It will try to grab either to print it. If neither is available, we will
just stick to printing the output of the query itself. It will improve
over time, so that it becomes clearer where it originated. At least for
the short version, we want to avoid printing the full query.


![image](https://user-images.githubusercontent.com/1307529/196852604-cc972660-cc9b-4d5f-abd1-0fe624f4ffb2.png)


Signed-off-by: Dominik Richter <dominik.richter@gmail.com>

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>