🐛 propagate errors about port-process-relationships (#418)
Previously when reading out the file descriptors, we would silently
ignore errors that happened there. However, it is quite common for
someone without sufficient permissions to read this list, e.g. when we
ask which port is opened by which process. When this happens, now we
propagate the errors and wrap it nicely so that users know it is a
permissions issue and can sudo their way out of it.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>