🧹 be more specific for gcp flags and use `--project-id` and `organization-id` (#470)
Before it was a bit unclear which arguments should be used. `project`
can also be the name of the project but we only supported the
project-id. This is a non-breaking change because the old flags stay
available but are hidden now. If the inventory is used, the old
properties are being supported as well.

```
cnquery shell gcp --organization-id 12345678
```