🧹 rename the gcp provider to google provider (#471)
after #470

This is no functional change nor a breaking change. CLI commands and
inventory settings stay the same. It is just the reflection that the
google api's reflect more than just GCP. Therefore we rename the
provider to be named google. The resource pack that provides access to
GCP resources stays the same. This allows us to use the same provider
for more google services going forward.