🏎️ switch llx.cache to internal + lightweight map
The difference to sync.Map is that it is going to be slower reading/writing, but more memory-efficient since no duplicate datastructures are kept around. Additionally, we can be type-safe and much more efficient with the key lookup, which should make up for some of this. The loss in potential performance is likely negligible, considering the majority of time is spent on I/O, not on data access.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>