🐛 prevent crash on empty builtin
When working with special resources (or in this case: dicts) that expand builtin functions, we sometimes run into the case where the builtin function does not exist. This was never checked, before the guardrail only looked for errors, but not if the function is actually non-nil. Remedy it to fix the panic.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>