🐛 prevent crash on dict[string].contains (#548)
Originally contains was only created for simple strings. However, we
have since started to support the keyword for arrays and maps. The
problem is: dicts can be either and the behavior betweenboth is very
different.

This PR adds the first level of support for contains on dicts, that
works on both arrays and strings. The printing is still shot, but this
is the first step to add proper support.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>