address PR feedback
Filter Secrets by Namespace (and use a map to keep track of the
available Secrets).

Rather than a combined list of certificates, break it up in a way that
follows the structure of the Ingress resource. So that each TLS entry
can hold the list of hosts (from the Ingress) and the list of
certificates found in the referenced Secret.

Signed-off-by: Joel Diaz <joel@mondoo.com>