🐛 add labels for vars (#648)
Note: This expands the proto because we offer a lower level interface
for users to interact. We don't want to magically move around metadata
that users of the API don't get access to. However, once completed the
additional data structures are cleared out, so the final object doesn't
grow in size.

Fixes https://github.com/mondoohq/cnquery/issues/574

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>